Cleanup Thermal node remains, clean up the example.
