+++ CMakeLists.txt
@@ -207,7 +207,6 @@ add_library(${TARGET_CORE_NAME}
     src/pipeline/node/XLinkOut.cpp
     src/pipeline/node/ColorCamera.cpp
     src/pipeline/node/Camera.cpp
-    src/pipeline/node/Thermal.cpp
     src/pipeline/node/ToF.cpp
     src/pipeline/node/MessageDemux.cpp
     src/pipeline/node/MonoCamera.cpp
