+++ include/depthai/pipeline/nodes.hpp
@@ -22,7 +22,6 @@
 #include "node/StereoDepth.hpp"
 #include "node/Sync.hpp"
 #include "node/SystemLogger.hpp"
-#include "node/Thermal.hpp"
 #include "node/ToF.hpp"
 #include "node/VideoEncoder.hpp"
 #include "node/Warp.hpp"
